+++ contracts/core/interfaces/IDelegatorPool.sol
@@ -6,5 +6,5 @@ import "./IERC677.sol";
 interface IDelegatorPool is IERC677 {
     function currentRate(address _token, uint16 _index) external view returns (uint256);
 
-    function stakeAllowance(address _account, uint _amount) external;
+    function stakeAllowance(address _account, uint256 _amount) external;
 }
