+++ contracts/core/tokens/LPLMigration.sol
@@ -11,7 +11,7 @@ import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 contract LPLMigration {
     using SafeERC20 for IERC20;
 
-    uint public constant MIGRATION_RATIO = 2;
+    uint256 public constant MIGRATION_RATIO = 2;
 
     address public lplToken;
     address public sdlToken;
